minor code improvements

+++ src/main/java/org/roaringbitmap/buffer/ImmutableRoaringBitmap.java
@@ -1412,7 +1412,7 @@ public class ImmutableRoaringBitmap
    */
   @Override
   public int[] toArray() {
-    final int[] array = new int[(int)this.getCardinality()];
+    final int[] array = new int[this.getCardinality()];
     int pos = 0, pos2 = 0;
     while (pos < this.highLowContainer.size()) {
       final int hs = BufferUtil.toIntUnsigned(this.highLowContainer.getKeyAtIndex(pos)) << 16;
